requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-groq==0.2.2
unstructured==0.16.11
rapidocr-onnxruntime==14.3
pymupdf==1.25.1
faiss-cpu
sentence-transformers==3.3.1
streamlit==1.41.1
# openai
# pdfminer.six
# pi-heif
# unstructured-inference

# For Faiss GPU -> conda install -c pytorch faiss-gpupi